+++ cyrus/lib/libcyr_cfg.h
@@ -40,7 +40,7 @@
  *
  */
 /*
- * $Id: libcyr_cfg.h,v 1.2.2.7 2005/02/20 03:23:45 shadow Exp $
+ * $Id: libcyr_cfg.h,v 1.2.2.8 2005/02/20 05:22:29 shadow Exp $
  */
 
 #ifndef INCLUDED_LIBCYR_CFG_H
@@ -106,8 +106,8 @@ enum cyrus_opt {
 
 union cyrus_config_value {
     const char *s; /* string */
-    int i; /* int */
-    int b; /* switch */
+    long i; /* int */
+    long b; /* switch */
 };
 
 struct cyrusopt_s {
